requirements.txt
langchain
langchain-google-genai
pandas
python-dotenv
